refactor(filemanager): remove unused structs and functions
