+++ lib/workload/stateless/filemanager/filemanager-http-lambda/src/main.rs
@@ -1,22 +1,15 @@
 use lambda_http::Error;
 use lambda_runtime::{run, service_fn, LambdaEvent};
-use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::util::SubscriberInitExt;
-use tracing_subscriber::{fmt, EnvFilter};
 
 use filemanager::clients::aws::s3::Client as S3Client;
 use filemanager::clients::aws::sqs::Client as SQSClient;
 use filemanager::database::Client as DbClient;
 use filemanager::handlers::aws::{create_database_pool, receive_and_ingest, update_credentials};
+use filemanager::handlers::init_tracing;
 
 #[tokio::main]
 async fn main() -> Result<(), Error> {
-    let env_filter = EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info"));
-
-    tracing_subscriber::registry()
-        .with(fmt::layer().json().without_time())
-        .with(env_filter)
-        .init();
+    init_tracing();
 
     let options = &create_database_pool().await?;
     run(service_fn(|_: LambdaEvent<()>| async move {
