+++ lib/workload/stateless/filemanager/filemanager-ingest-lambda/Cargo.toml
@@ -10,7 +10,6 @@ rust-version.workspace = true
 [dependencies]
 tokio = { version = "1", features = ["macros"] }
 tracing = { version = "0.1", features = ["log"] }
-tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt"] }
 
 aws_lambda_events = "0.15"
 lambda_runtime = "0.11"
