+++ lib/workload/stateless/filemanager/filemanager/Cargo.toml
@@ -20,9 +20,9 @@ async-trait = "0.1"
 futures = "0.3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
 tracing = "0.1"
+tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt"] }
 
 chrono = { version = "0.4", features = ["serde"] }
-dotenvy = "0.15"
 thiserror = "1"
 uuid = { version = "1", features = ["v7"] }
 mockall = "0.12"
