+++ lib/workload/stateless/filemanager/filemanager/src/database/aws/migration.rs
@@ -1,3 +1,6 @@
+//! Database migration logic.
+//!
+
 use async_trait::async_trait;
 use sqlx::migrate;
 use sqlx::migrate::Migrator;
