+++ lib/workload/stateless/filemanager/filemanager/src/error.rs
@@ -1,4 +1,4 @@
-//! Errors used by the crate.
+//! Errors used by the filemanager crate.
 //!
 
 use std::result;
@@ -11,36 +11,18 @@ pub type Result<T> = result::Result<T, Error>;
 /// Error types for the filemanager.
 #[derive(Error, Debug)]
 pub enum Error {
-    /// File not found by id.
-    #[error("file not found: `{0}`")]
-    NotFound(String),
-    /// File operation unauthorized
-    #[error("unauthorized: `{0}`")]
-    Unauthorized(String),
     #[error("SQL error: `{0}`")]
     SQLError(String),
     #[error("SQL migrate error: `{0}`")]
     MigrateError(String),
-    #[error("SQS client error: `{0}`")]
-    SQSClientError(String),
-    #[error("SQS message receive error: `{0}`")]
-    SQSReceiveError(String),
+    #[error("SQS error: `{0}`")]
+    SQSError(String),
     #[error("deserialization error: `{0}`")]
     DeserializeError(String),
-    #[error("Db client error: `{0}`")]
-    DbClientError(String),
-    #[error("Missing SQS URL: `{0}`")]
-    MissingSQSUrl(String),
     #[error("Missing environment variable: `{0}`")]
     MissingEnvironmentVariable(String),
-    #[error("S3 object not found: `{0}`")]
-    S3ObjectNotFound(String),
     #[error("S3 error: `{0}`")]
     S3Error(String),
-    #[error("Config error: `{0}`")]
-    ConfigError(String),
-    #[error("Ingester error: `{0}`")]
-    IngesterError(String),
     #[error("credential generator error: `{0}`")]
     CredentialGeneratorError(String),
 }
