+++ lib/workload/stateless/filemanager/filemanager/src/handlers/mod.rs
@@ -1,4 +1,18 @@
 //! This module contains event handlers for filemanager functionality.
 //!
 
+use lambda_runtime::tracing::subscriber::{fmt, EnvFilter};
+use tracing_subscriber::layer::SubscriberExt;
+use tracing_subscriber::util::SubscriberInitExt;
+
 pub mod aws;
+
+/// Initialize tracing for application code.
+pub fn init_tracing() {
+    let env_filter = EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info"));
+
+    tracing_subscriber::registry()
+        .with(fmt::layer().json().without_time())
+        .with(env_filter)
+        .init();
+}
