--- lib/workload/stateless/filemanager/scripts/watch.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-cargo watch -w *.toml -w *.rs -- ./scripts/deploy.sh
